Initial commit
I started working on the database before I watched the rest of the videos so my code was all messed up.  I had to take it all back and for some reason my xml folder and pref xml file were missing and so was the arrays for the preference categories and sort options.  Starting from scratch at end of P1